Cleanned up a couple of helper files.
