+++ src/myoddweb.directorywatcher/utils/Helper/NativeLibrary.cs
@@ -1,0 +1,14 @@
+﻿namespace myoddweb.directorywatcher.utils.Helper
+{
+  internal static class NativeLibrary
+  {
+    [DllImport("kernel32.dll")]
+    public static extern IntPtr LoadLibrary(string dllToLoad);
+
+    [DllImport("kernel32.dll")]
+    public static extern IntPtr GetProcAddress(IntPtr hModule, string procedureName);
+
+    [DllImport("kernel32.dll")]
+    public static extern bool FreeLibrary(IntPtr hModule);
+  }
+}
